modularizando providers
