+++ lib/presentation/features/create/screens/create_student_module.dart
@@ -3,7 +3,7 @@ import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:nawiapp/domain/models/student/entity/student.dart';
 import 'package:nawiapp/domain/models/student/entity/student_age.dart';
 import 'package:nawiapp/domain/models/student/summary/student_summary.dart';
-import 'package:nawiapp/presentation/features/create/providers/create_student_form_provider.dart';
+import 'package:nawiapp/presentation/features/create/providers/student/create_student_form_provider.dart';
 import 'package:nawiapp/presentation/shared/submit_status.dart';
 import 'package:nawiapp/presentation/widgets/another_student_element.dart';
 import 'package:nawiapp/presentation/widgets/loading_process_button.dart';
