+++ lib/presentation/widgets/another_student_element.dart
@@ -5,8 +5,8 @@ import 'package:nawiapp/data/mappers/student_mapper.dart';
 import 'package:nawiapp/domain/models/student/entity/student.dart';
 import 'package:nawiapp/domain/models/student/summary/student_summary.dart';
 import 'package:nawiapp/domain/services/student_service_base.dart';
-import 'package:nawiapp/presentation/features/create/providers/initial_student_form_data_provider.dart';
-import 'package:nawiapp/presentation/features/create/providers/selectable_element_for_create_provider.dart';
+import 'package:nawiapp/presentation/features/create/providers/student/initial_student_form_data_provider.dart';
+import 'package:nawiapp/presentation/features/create/providers/student/selectable_element_for_create_provider.dart';
 import 'package:nawiapp/presentation/features/home/extra/menu_tabs.dart';
 import 'package:nawiapp/presentation/features/home/providers/tab_index_provider.dart';
 import 'package:nawiapp/presentation/features/search/providers/general_loading_search_student_provider.dart';
